fix warnings

+++ test/test_task_cifar10.cpp
@@ -63,7 +63,7 @@ NANO_CASE(construction)
         const size_t max_duplicates = 0;
         NANO_CHECK_LESS_EQUAL(nano::check_duplicates(*task), max_duplicates);
         NANO_CHECK_LESS_EQUAL(nano::check_intersection(*task), max_duplicates);
-        NANO_CHECK_EQUAL(labels.size(), odims.size());
+        NANO_CHECK_EQUAL(labels.size(), static_cast<size_t>(odims.size()));
 }
 
 NANO_END_MODULE()
